refactorizando los servicios y agregando los querymethods en los repositorios

+++ src/main/java/com/besysoft/practica/services/interfaces/GeneroService.java
@@ -1,13 +1,13 @@
 package com.besysoft.practica.services.interfaces;
 
-import com.besysoft.practica.dominio.Genero;
+import com.besysoft.practica.entities.Genero;
 
 import java.util.Optional;
 
 public interface GeneroService {
 
     Iterable<Genero>obtenerTodos() throws Exception;
-    Optional<Genero>buscarPorId(int id) throws Exception;
+    Optional<Genero>buscarPorId(Long id) throws Exception;
 
     Optional<Genero>buscarPorNombre(String nombre) throws Exception;
